ChaosMod: Reincorporate mission fail prevention into 'Fake Teleport'
This time instead of spoofing all entity coords using a GET_ENTITY_COORDS hook we are preventing the execution of any script threads (with the exception of main and main_persistent which are essential for Scripthook), thanks to SeanMirrsen for the idea
