--- ChaosMod/Memory/Hooks/EntityCoordsHook.h
@@ -1,10 +1,0 @@
-#pragma once
-
-struct Vector3;
-
-namespace Hooks
-{
-	void EnableFakeTpHook();
-
-	void DisableFakeTpHook();
-}
