+++ ChaosMod/Memory/Hooks/ScriptThreadRunHook.cpp
@@ -1,0 +1,51 @@
+#include <stdafx.h>
+
+static bool m_enabledHook = false;
+
+__int64(*OG_rage__scrThread__Run)(__int64 a1);
+__int64 HK_rage__scrThread__Run(__int64 a1)
+{
+	if (m_enabledHook)
+	{
+		const char* scriptName = reinterpret_cast<const char*>(a1 + 0xD0);
+
+		// ScriptHook relies on these to run our script thread
+		// We don't want to block ourselves of course :p
+		if (strcmp(scriptName, "main") && strcmp(scriptName, "main_persistent"))
+		{
+			return 0;
+		}
+	}
+
+	return OG_rage__scrThread__Run(a1);
+}
+
+static bool OnHook()
+{
+	Handle handle;
+
+	handle = Memory::FindPattern("48 89 5C 24 08 48 89 6C 24 10 48 89 74 24 18 57 41 56 41 57 48 83 EC 20 48 8D 81 D0 00 00 00");
+	if (!handle.IsValid())
+	{
+		return false;
+	}
+
+	Memory::AddHook(handle.Get<void>(), HK_rage__scrThread__Run, &OG_rage__scrThread__Run);
+
+	return true;
+}
+
+static RegisterHook registerHook(OnHook, "rage::scrThread::Run");
+
+namespace Hooks
+{
+	void EnableScriptThreadBlock()
+	{
+		m_enabledHook = true;
+	}
+
+	void DisableScriptThreadBlock()
+	{
+		m_enabledHook = false;
+	}
+}
