+++ ChaosMod/Memory/Hooks/ScriptThreadRunHook.h
@@ -1,0 +1,10 @@
+#pragma once
+
+struct Vector3;
+
+namespace Hooks
+{
+	void EnableScriptThreadBlock();
+
+	void DisableScriptThreadBlock();
+}
